requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
babel==2.4.0
dj-database-url==0.4.2
django-authtools==1.5.0
django-debug-toolbar==1.7
django-dotenv==1.4.1
django-enumfields==0.9.0
django==1.11
markdown==2.6.8
psycopg2==2.7.1
pytz==2017.2              # via babel, django
sqlparse==0.2.3           # via django-debug-toolbar
